+++ pdp/portals/mbcn_cmip6.py
@@ -1,8 +1,7 @@
-'''This portal serves the CMIP6 data downscaled by BCCAQv2
-for all Canada. The UI is similar to the CMIP5 BCCAQv2 data, and
-the use the same map component (canada_ex_map.js), but 
-different frontend controllers (cmip6_bccaq2_app.js). It also contains
-additional headings for the PCIC12 models.
+'''This portal serves the CMIP6 data downscaled by MBCn
+for all Canada. The UI is similar to the CMIP6 BCCAQv2 data, and
+the use the same map component (canada_ex_map.js), but
+different frontend controllers (cmip6_mbcn_app.js).
 '''
 from pdp.portals import make_raster_frontend, data_server
 from pdp_util.ensemble_members import EnsembleMemberLister
@@ -12,9 +11,9 @@ import re
 __all__ = ['url_base', 'mk_frontend', 'mk_backend']
 
 
-ensemble_name = 'bccaq2_cmip6'
-url_base = '/downscaled_cmip6'
-title = 'Canadian Downscaled Climate Scenarios - Univariate (CMIP6): CanDCS-U6'
+ensemble_name = 'mbcn_cmip6'
+url_base = '/downscaled_cmip6_multi'
+title = 'Canadian Downscaled Climate Scenarios - Multivariate (CMIP6): CanDCS-M6'
 
 
 class CMIP6EnsembleLister(EnsembleMemberLister):
@@ -61,7 +60,7 @@ def mk_frontend(config):
     return make_raster_frontend(config, ensemble_name, url_base,
                                 title, CMIP6EnsembleLister,
                                  ['js/canada_ex_map.js',
-                                 'js/cmip6_bccaq2_app.js'])
+                                 'js/cmip6_mbcn_app.js'])
  
 
 def mk_backend(config):